favourites context established
moved favourites functionality into context so other components of the app can access
